Address Jungtaek's comments
